Implement internal, default, and collector APIs
