+++ src/default.rs
@@ -1,20 +1,20 @@
-//! The default collector for garbage collection.
+//! The default garbage collector.
 //!
-//! For each thread, a handle is lazily initialized on its first use, when the current thread is
-//! registered in the default collector.  If initialized, the thread's handle will get destructed
-//! on thread exit, which in turn unregisters the thread.
+//! For each thread, a participant is lazily initialized on its first use, when the current thread
+//! is registered in the default collector.  If initialized, the thread's participant will get
+//! destructed on thread exit, which in turn unregisters the thread.
 
-use collector::Global;
-use handle::{Handle, Scope};
+use internal::{Global, Participant};
+use scope::Scope;
 
 lazy_static! {
-    /// The default global data.
+    /// The global data for the default garbage collector.
     static ref COLLECTOR: Global = Global::new();
 }
 
 thread_local! {
-    /// The thread-local handle for the default global data.
-    static HANDLE: Handle<&'static Global> = Handle::new(&COLLECTOR.as_ref());
+    /// The per-thread participant for the default garbage collector.
+    static PARTICIPANT: Participant = Participant::new(&COLLECTOR);
 }
 
 /// Pin the current thread.
@@ -23,13 +23,13 @@ where
     F: for<'scope> FnOnce(Scope<'scope>) -> R,
 {
     // FIXME(jeehoonkang): thread-local storage may be destructed at the time `pin()` is called. For
-    // that case, we should use `HANDLE.try_with()` instead.
-    HANDLE.with(|handle| handle.pin(f))
+    // that case, we should use `PARTICIPANT.try_with()` instead.
+    PARTICIPANT.with(|participant| participant.pin(&COLLECTOR, f))
 }
 
 /// Check if the current thread is pinned.
 pub fn is_pinned() -> bool {
     // FIXME(jeehoonkang): thread-local storage may be destructed at the time `pin()` is called. For
-    // that case, we should use `HANDLE.try_with()` instead.
-    HANDLE.with(|handle| handle.is_pinned())
+    // that case, we should use `PARTICIPANT.try_with()` instead.
+    PARTICIPANT.with(|participant| participant.is_pinned())
 }
